# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample documentation for the SHA project (FrodidApp)",
    "version": "1.0.0",
    "title": "FrodidApp",
    "termsOfService": "",
    "contact": {
      "email": "semerkhet40@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ec2-18-185-66-15.eu-central-1.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/visits": {
      "post": {
        "tags": [
          "visits"
        ],
        "summary": "Creates new visit",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Visits"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": [
          "visits"
        ],
        "summary": "Gets a list of all visits",
        "description": "This endpoint uses a get request to retrieve all visits",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "items_per_page",
            "type": "integer",
            "required": false
          },
          {
            "in": "query",
            "name": "page_number",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/visits/{id}": {
      "get": {
        "tags": [
          "visits"
        ],
        "summary": "Gets a single visit",
        "description": "This endpoint uses a get request to retrieve a specific visit",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User unique identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Visits"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "visits"
        ],
        "summary": "Updates a user in the visits table",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of visit that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created visit object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Visits"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "visits"
        ],
        "summary": "Deletes a visit",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Visit id to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Visit not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Visits": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time",
          "example": "2000-01-31"
        },
        "comment": {
          "type": "string",
          "example": "Patient requires more medicine and a second visit"
        },
        "start_time": {
          "type": "string",
          "example": "03:07:00"
        },
        "end_time": {
          "type": "string",
          "example": "05:10:00"
        },
        "user_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
  },
...
}